test e2e

+++ src/controller/users.js
@@ -1,5 +1,7 @@
 const User = require('../models/user');
-const { pagination, validateUser } = require('../utils/utils');
+const {
+  pagination, validateUser, isAValidEmail, isAWeakPassword,
+} = require('../utils/utils');
 const { isAdmin } = require('../middleware/auth');
 
 // GET '/users'
@@ -27,8 +29,6 @@ const getUsers = async (req, res, next) => {
 
 const getOneUser = async (req, res, next) => {
   try {
-    console.log('HOLAAAAAAAAAAAAAAAAAAAAAAAAAAAAA');
-    console.log(req.authToken);
     const { uid } = req.params;
     const value = validateUser(uid);
     const user = await User.findOne(value).lean();
@@ -39,7 +39,6 @@ const getOneUser = async (req, res, next) => {
     if (req.authToken.uid === user._id.toString() || isAdmin(req)) return res.json(user);
     return next(403);
   } catch (err) {
-    console.log(err);
     return next(err);
   }
 };
@@ -48,19 +47,26 @@ const getOneUser = async (req, res, next) => {
 
 const newUser = async (req, res, next) => {
   try {
+    const { email, password } = req.body;
+
+    if (!email || !password) {
+      return next(400);
+    }
+
+    if (isAWeakPassword(password) || !isAValidEmail(email)) return next(400);
+
     const findUser = await User.findOne({ email: req.body.email });
 
     if (findUser) {
-      return res.status(409).json({
+      return res.status(403).json({
         message: '(Error) El usuario ya se encuentra registrado',
       });
     }
 
     const newUser = new User(req.body);
     const userSaved = await newUser.save(newUser);
-    const user = await User.findOne({ _id: userSaved._id });
-
-    res.status(200).json(user);
+    const user = await User.findOne({ _id: userSaved._id }).select('-password');
+    return res.status(200).json(user);
   } catch (err) {
     console.info(err);
     next(err);
@@ -71,14 +77,46 @@ const newUser = async (req, res, next) => {
 
 const updateUser = async (req, res, next) => {
   try {
+    const { uid } = req.params;
+    console.log(req.body);
+
+    const value = validateUser(uid);
+    const user = await User.findOne(value);
+
+    const userId = user._id.toString() || '';
+
+    console.log(req.authToken.uid !== userId);
+    console.log(isAdmin(req));
+    console.info('LINEA 1', req.body);
+    console.info(req.authToken);
+
+    if (req.authToken.uid !== userId && !isAdmin(req)) return next(403);
+    console.info('LINEA 1-2', req.body);
+
+    console.log(!isAdmin(req));
+    console.log(!isAdmin(req) && user.roles);
+
+    if (!isAdmin(req) && user.roles) return next(403);
+
+    console.info('LINEA 2', req.body);
+    console.info(req.authToken);
+    if (Object.keys(req.body).length === 0) return next(403);
+
+    console.info('LINEA 3', req.body);
+    console.info(req.authToken);
+    if (!user) return next(404);
+
     const userUpdate = await User.findOneAndUpdate(
-      { _id: req.params.uid },
+      { value },
       { $set: req.body },
       { new: true, useFindAndModify: false },
     ); // .select('-__v');
-    res.status(200).json(userUpdate);
+
+    return res.status(200).json(userUpdate);
   } catch (err) {
-    next(err);
+    console.info('LINEA 4', req.body);
+    console.info(req.authToken);
+    next(404);
   }
 };
 
@@ -86,14 +124,21 @@ const updateUser = async (req, res, next) => {
 
 const deleteOneUser = async (req, res, next) => {
   try {
-    const userDeleted = await User.findOne({ _id: req.params.uid });
+    const { uid } = req.params;
+    const value = validateUser(uid);
+    const userDeleted = await User.findOne(value);
+
+    const userId = userDeleted._id || '';
+
+    if (!userDeleted) return next(404);
+
+    if (req.authToken.uid !== userId || !isAdmin(req)) return next(403);
+
     await User.findByIdAndDelete({ _id: req.params.uid });
-    if (userDeleted) {
-      res.status(200).json(userDeleted);
-    }
-    res.status(400).json({ message: 'Usuario a eliminar no existe' });
+
+    res.status(200).json(userDeleted);
   } catch (err) {
-    next(err);
+    next(404);
   }
 };
